requirements.txt
psycopg2-binary
python-dotenv
alembic
python-dateutil
fastapi
uvicorn
pytest